Backend/requirements.txt
flask
pytest
pytest-flask
psycopg2-binary